+++ m5-boat-display/m5-draw.h
@@ -3,6 +3,6 @@
 
 void fillArc(int x, int y, int start_angle, int seg_count, int rx, int ry, int w, unsigned int color);
 
-float degToRad(int angleDeg);
+float degToRad(float angleDeg);
 
 #endif // __m5_draw_h
